magento-engcom/community-portal#78 Apply the new mockup from the UI team - moved the voteCount next to the likeButton - wrap the addProjectButton with withAuth HOC to manage user access
